+++ packages/hightide-design/src/tokens/mappings/to-components.ts
@@ -1,8 +1,8 @@
-import type { ComponentColorTokens } from '../../../types/component-colors'
+import type { ComponentColorTokens } from '../../types/component-colors'
 import type {
   HightideColorPalleteTokens,
   HightideSemanticColorTokens
-} from '../../../types/hightide'
+} from '../../types/hightide'
 
 export type ToComponentsArgs<PrimitiveTokens, SemanticTokens> = {
   themeName: string,