+++ packages/hightide-design/src/tokens/themes/index.ts
@@ -1,4 +1,3 @@
 export * from './dark'
 export * from './light'
-export * from './mappings'
 export * from './themes'