+++ packages/hightide-design/src/tokens/themes/light.ts
@@ -1,14 +1,14 @@
 import { colorPalettes } from '../color-palettes'
-import { constructTheme } from '../../hooks/constructTheme'
-import type { HightideDesignTokens } from '../../types/hightide'
-import { toHightideComponentTokens } from './mappings/to-components'
-import { toHightideSemanticTokens } from './mappings/to-semantic'
-import { toHightideTheme } from './mappings/to-theme'
+import { constructThemeTokens } from '../../utils/constructThemeTokens'
+import type { HightideThemeTokens } from '../../types/hightide'
+import { toHightideComponentTokens } from '../mappings/to-components'
+import { toHightideSemanticTokens } from '../mappings/to-semantic'
+import { toHightideTheme } from '../mappings/to-theme'
 
-export const lightTheme = constructTheme({
+export const lightTheme = constructThemeTokens({
   themeName: 'light',
   primitiveTokens: colorPalettes,
   toSemantic: toHightideSemanticTokens,
   toComponents: toHightideComponentTokens,
   toTheme: toHightideTheme,
-}) satisfies HightideDesignTokens
+}) satisfies HightideThemeTokens