+++ packages/hightide-design/src/tokens/themes/themes.ts
@@ -1,8 +1,8 @@
-import type { HightideDesignTokens } from '../../types/hightide'
+import type { HightideThemeTokens } from '../../types/hightide'
 import { darkTheme } from './dark'
 import { lightTheme } from './light'
 
-export const Themes = {
+export const ThemeTokens = {
   light: lightTheme,
   dark: darkTheme,
-} as const satisfies Record<string, HightideDesignTokens>
+} as const satisfies Record<string, HightideThemeTokens>