+++ packages/hightide-native/src/components/user-interaction/IconButton.tsx
@@ -10,7 +10,7 @@ import {
 } from 'react-native'
 import type { LucideIcon } from 'lucide-react-native'
 
-import type { ColoringType } from '@helpwave/hightide-design/helpers'
+import type { ColoringType } from '@helpwave/hightide-design/utils'
 import type {
   ButtonColoringStyle,
   ElementSize