+++ packages/hightide-native/src/theme/resolvers/input.ts
@@ -3,7 +3,7 @@ import type { TextStyle } from 'react-native'
 import { componentLayouts } from '@helpwave/hightide-design/tokens'
 import type {
   ComponentColorTokens,
-  HightideDesignTokens as DesignTokensTheme
+  HightideThemeTokens as DesignTokensTheme
 } from '@helpwave/hightide-design/types'
 
 import type { HightideSemanticColors } from '../types/color'