+++ packages/hightide-native/stories/Theme/ThemeExtension.stories.tsx
@@ -3,21 +3,21 @@ import type {
   Meta,
   StoryObj
 } from '@storybook/react-native-web-vite'
-import { constructTheme } from '@helpwave/hightide-design/hooks'
+import { constructThemeTokens } from '@helpwave/hightide-design/utils'
 import {
   colorPalettes,
   toHightideComponentTokens,
   toHightideSemanticTokens,
   toHightideTheme
 } from '@helpwave/hightide-design/tokens'
-import type { HightideDesignTokens } from '@helpwave/hightide-design/types'
+import type { HightideThemeTokens } from '@helpwave/hightide-design/types'
 
 import { Button } from '../../src/components/user-interaction/Button'
 import { Chip } from '../../src/components/visualization-and-display/Chip'
 import { HightideConfigUtils } from '../../src/global-contexts/hightide-config/HightideConfigUtils'
 import { useTheme } from '../../src/global-contexts/theme/ThemeContext'
 import { ThemeProvider } from '../../src/global-contexts/theme/ThemeProvider'
-import { createTheme } from '../../src/theme/themes/createTheme'
+import { createHightideTheme } from '../../src/theme/themes/createHightideTheme'
 import {
   ThemeSelect,
   ThemeStoryFrame
@@ -41,15 +41,15 @@ const toOceanSemantic: typeof toHightideSemanticTokens = (args) => {
   }
 }
 
-const oceanDesignTokens = constructTheme({
+const oceanDesignTokens = constructThemeTokens({
   themeName: 'ocean',
   primitiveTokens: colorPalettes,
   toSemantic: toOceanSemantic,
   toComponents: toHightideComponentTokens,
   toTheme: toHightideTheme,
-}) satisfies HightideDesignTokens
+}) satisfies HightideThemeTokens
 
-const oceanTheme = createTheme(oceanDesignTokens)
+const oceanTheme = createHightideTheme(oceanDesignTokens)
 
 const oceanSupportedThemes = {
   ...HightideConfigUtils.defaultSupportedThemes,